Webapplication interface cleanup
